+++ recbole/data/dataset/dataset.py
@@ -650,7 +650,7 @@ class Dataset(torch.utils.data.Dataset):
                 elif ftype == FeatureType.FLOAT:
                     feat[field].fillna(value=feat[field].mean(), inplace=True)
                 else:
-                    dtype = np.int32 if ftype == FeatureType.TOKEN_SEQ else np.float
+                    dtype = np.int64 if ftype == FeatureType.TOKEN_SEQ else np.float
                     feat[field] = feat[field].apply(
                         lambda x: np.array([], dtype=dtype)
                         if isinstance(x, float)
@@ -2018,7 +2018,7 @@ class Dataset(torch.utils.data.Dataset):
             row_num, max_col_num = self.item_num, self.user_num
             row_ids, col_ids = item_ids, user_ids
 
-        history_len = np.zeros(row_num, dtype=np.int32)
+        history_len = np.zeros(row_num, dtype=np.int64)
         for row_id in row_ids:
             history_len[row_id] += 1
 
@@ -2034,7 +2034,7 @@ class Dataset(torch.utils.data.Dataset):
                 f"{col_num / max_col_num * 100}% of the total."
             )
 
-        history_matrix = np.zeros((row_num, col_num), dtype=np.int32)
+        history_matrix = np.zeros((row_num, col_num), dtype=np.int64)
         history_value = np.zeros((row_num, col_num))
         history_len[:] = 0
         for row_id, value, col_id in zip(row_ids, values, col_ids):
